[Feat] #254 - 통계 메뉴 발주/판매 분리 및 페이지 분할
- 통계 메뉴를 드롭다운 그룹(발주 통계 / 판매·매출 분석)으로 분리
- StatisticsView를 발주 통계 전용으로 리팩터링
- SalesStatisticsView 신규 추가 (가맹점·제품·카테고리별 매출 분석)
- 공유 데이터 생성 로직을 utils/orderData.js로 분리
- /statistics/sales 라우트 추가
